move summary network fixtures into `test_networks/conftest.py`
